requirements.txt
fastapi
uvicorn[standard]
feedparser
httpx
aiosqlite
apscheduler
PyYAML
python-dateutil
